This PR improves the optimized single server get multi, the primary fixes were reducing the number of chomps!, reducing allocations with byteslice, preallocate expected hash size...


benchmark before:

```
Comparison:
       Ecc get_multi:      453.6 i/s
 EccClient get_multi:      448.8 i/s - same-ish: difference falls within error
      Sock get multi:      431.3 i/s - same-ish: difference falls within error
  EccStore get_multi:      424.1 i/s - same-ish: difference falls within error
DalliClient get_multi:      366.6 i/s - 1.24x  slower
DalliStore get_multi:      360.3 i/s - 1.26x  slower
     Dalli get multi:      344.7 i/s - 1.32x  slower
 FileStore get multi:      341.7 i/s - 1.33x  slower
MemCacheStore get multi:      327.6 i/s - 1.38x  slower
```

benchmark after

```
Comparison:
       Ecc get_multi:      424.2 i/s
      Sock get multi:      408.5 i/s - same-ish: difference falls within error
DalliStore get_multi:      392.2 i/s - same-ish: difference falls within error
 EccClient get_multi:      377.6 i/s - same-ish: difference falls within error
DalliClient get_multi:      373.5 i/s - same-ish: difference falls within error
  EccStore get_multi:      356.1 i/s - same-ish: difference falls within error
     Dalli get multi:      349.4 i/s - 1.21x  slower
```